# swagger/swagger.json
{"swagger":"2.0","info":{"title":"goa study chat","description":"goa study chat api","version":""},"host":"localhost:9089","basePath":"/api","schemes":["http"],"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/accounts":{"get":{"tags":["account"],"summary":"show account","description":"Retrieve account with given id or something","operationId":"account#show","produces":["application/vnd.goa.error","application/vnd.account+json; type=collection"],"parameters":[{"name":"ids","in":"query","required":false,"type":"array","items":{"type":"string"},"collectionFormat":"multi"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/AccountCollection"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http"]}},"/accounts/me":{"get":{"tags":["account"],"summary":"showMe account","description":"Retrieve my account\n\nRequired security scopes:\n  * `api:access`","operationId":"account#showMe","produces":["application/vnd.goa.error","application/vnd.account+json"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Account"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http"],"security":[{"jwt":["api:access"]}]}},"/rooms":{"get":{"tags":["room"],"summary":"list room","description":"Retrieve all rooms.","operationId":"room#list","produces":["application/vnd.room+json; type=collection"],"parameters":[{"name":"limit","in":"query","required":false,"type":"integer"},{"name":"offset","in":"query","required":false,"type":"integer"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RoomCollection"}},"404":{"description":"Not Found"}},"schemes":["http"]},"post":{"tags":["room"],"summary":"post room","description":"Create new Room\n\nRequired security scopes:\n  * `api:access`","operationId":"room#post","produces":["application/vnd.goa.error"],"parameters":[{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/RoomPayload"}}],"responses":{"201":{"description":"Resource created","headers":{"Location":{"description":"href to created resource","type":"string","pattern":"/rooms/[0-9]+"}}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"401":{"description":"Unauthorized"}},"schemes":["http"],"security":[{"jwt":["api:access"]}]}},"/rooms/{name}":{"get":{"tags":["room"],"summary":"show room","description":"Retrieve room with given name","operationId":"room#show","produces":["application/vnd.goa.error","application/vnd.room+json"],"parameters":[{"name":"name","in":"path","description":"Name of room","required":true,"type":"string","maxLength":20,"minLength":3,"pattern":"[a-z|0-9]+"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Room"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http"]}},"/rooms/{name}/messages":{"get":{"tags":["message"],"summary":"list message","description":"Retrieve all messages.","operationId":"message#list","produces":["application/vnd.response_messages+json"],"parameters":[{"name":"name","in":"path","description":"Name of room","required":true,"type":"string","maxLength":20,"minLength":3,"pattern":"[a-z|0-9]+"},{"name":"nextCursor","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Response_messages"}},"404":{"description":"Not Found"}},"schemes":["http"]},"post":{"tags":["message"],"summary":"post message","description":"Create new message\n\nRequired security scopes:\n  * `api:access`","operationId":"message#post","parameters":[{"name":"name","in":"path","description":"Name of room","required":true,"type":"string","maxLength":20,"minLength":3,"pattern":"[a-z|0-9]+"},{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/MessagePayload"}}],"responses":{"201":{"description":"Resource created","headers":{"Location":{"description":"href to created resource","type":"string","pattern":"^/rooms/[0-9]+/messages/[0-9]+$"}}},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"}},"schemes":["http"],"security":[{"jwt":["api:access"]}]}},"/rooms/{name}/messages/{messageID}":{"get":{"tags":["message"],"summary":"show message","description":"Retrieve message with given id","operationId":"message#show","produces":["application/vnd.goa.error","application/vnd.message+json"],"parameters":[{"name":"messageID","in":"path","required":true,"type":"integer"},{"name":"name","in":"path","description":"Name of room","required":true,"type":"string","maxLength":20,"minLength":3,"pattern":"[a-z|0-9]+"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Message"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http"]}}},"definitions":{"Account":{"title":"Mediatype identifier: application/vnd.account+json; view=default","type":"object","properties":{"googleUserID":{"type":"string","example":"Et iure accusamus doloremque qui magnam assumenda."},"lastLogin":{"type":"string","example":"2012-02-25T05:57:31Z","format":"date-time"},"name":{"type":"string","example":"In voluptas placeat est."},"picture":{"type":"string","example":"Occaecati explicabo eos ipsa dolorem accusantium."}},"description":"A account (default view)","example":{"googleUserID":"Et iure accusamus doloremque qui magnam assumenda.","lastLogin":"2012-02-25T05:57:31Z","name":"In voluptas placeat est.","picture":"Occaecati explicabo eos ipsa dolorem accusantium."},"required":["googleUserID","picture","lastLogin"]},"AccountCollection":{"title":"Mediatype identifier: application/vnd.account+json; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/Account"},"description":"AccountCollection is the media type for an array of Account (default view)","example":[{"googleUserID":"Et iure accusamus doloremque qui magnam assumenda.","lastLogin":"2012-02-25T05:57:31Z","name":"In voluptas placeat est.","picture":"Occaecati explicabo eos ipsa dolorem accusantium."},{"googleUserID":"Et iure accusamus doloremque qui magnam assumenda.","lastLogin":"2012-02-25T05:57:31Z","name":"In voluptas placeat est.","picture":"Occaecati explicabo eos ipsa dolorem accusantium."},{"googleUserID":"Et iure accusamus doloremque qui magnam assumenda.","lastLogin":"2012-02-25T05:57:31Z","name":"In voluptas placeat est.","picture":"Occaecati explicabo eos ipsa dolorem accusantium."}]},"Message":{"title":"Mediatype identifier: application/vnd.message+json; view=default","type":"object","properties":{"auther":{"type":"string","example":"Aut laboriosam corporis ducimus beatae occaecati temporibus."},"content":{"type":"string","example":"r4jwj9vhp","minLength":1,"maxLength":400},"created":{"type":"string","example":"2013-09-21T07:19:10Z","format":"date-time"},"id":{"type":"integer","example":5176058875280461894,"format":"int64"}},"description":"A Message (default view)","example":{"auther":"Aut laboriosam corporis ducimus beatae occaecati temporibus.","content":"r4jwj9vhp","created":"2013-09-21T07:19:10Z","id":5176058875280461894},"required":["id","content","auther","created"]},"MessageCollection":{"title":"Mediatype identifier: application/vnd.message+json; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/Message"},"description":"MessageCollection is the media type for an array of Message (default view)","example":[{"auther":"Aut laboriosam corporis ducimus beatae occaecati temporibus.","content":"r4jwj9vhp","created":"2013-09-21T07:19:10Z","id":5176058875280461894},{"auther":"Aut laboriosam corporis ducimus beatae occaecati temporibus.","content":"r4jwj9vhp","created":"2013-09-21T07:19:10Z","id":5176058875280461894},{"auther":"Aut laboriosam corporis ducimus beatae occaecati temporibus.","content":"r4jwj9vhp","created":"2013-09-21T07:19:10Z","id":5176058875280461894}]},"MessagePayload":{"title":"MessagePayload","type":"object","properties":{"auther":{"type":"string","example":"Quaerat totam voluptatem qui voluptatem eum."},"content":{"type":"string","example":"sp","minLength":1,"maxLength":400},"created":{"type":"string","example":"1992-10-18T18:44:42Z","format":"date-time"}},"example":{"auther":"Quaerat totam voluptatem qui voluptatem eum.","content":"sp","created":"1992-10-18T18:44:42Z"},"required":["content","auther","created"]},"Response_messages":{"title":"Mediatype identifier: application/vnd.response_messages+json; view=default","type":"object","properties":{"messages":{"$ref":"#/definitions/MessageCollection"},"next":{"type":"string","example":"Voluptatem ad quas quas doloribus repudiandae tempora."}},"description":"A ResponseMessage (default view)","example":{"messages":[{"auther":"Aut laboriosam corporis ducimus beatae occaecati temporibus.","content":"r4jwj9vhp","created":"2013-09-21T07:19:10Z","id":5176058875280461894},{"auther":"Aut laboriosam corporis ducimus beatae occaecati temporibus.","content":"r4jwj9vhp","created":"2013-09-21T07:19:10Z","id":5176058875280461894},{"auther":"Aut laboriosam corporis ducimus beatae occaecati temporibus.","content":"r4jwj9vhp","created":"2013-09-21T07:19:10Z","id":5176058875280461894}],"next":"Voluptatem ad quas quas doloribus repudiandae tempora."},"required":["messages","next"]},"Room":{"title":"Mediatype identifier: application/vnd.room+json; view=default","type":"object","properties":{"created":{"type":"string","description":"Date of creation","example":"2015-08-11T06:54:39Z","format":"date-time"},"description":{"type":"string","description":"description of room","example":"room description","maxLength":400},"name":{"type":"string","description":"Name of room","example":"room001","pattern":"[a-z|0-9]+","minLength":3,"maxLength":20}},"description":"A room (default view)","example":{"created":"2015-08-11T06:54:39Z","description":"room description","name":"room001"},"required":["name","description","created"]},"RoomCollection":{"title":"Mediatype identifier: application/vnd.room+json; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/Room"},"description":"RoomCollection is the media type for an array of Room (default view)","example":[{"created":"2015-08-11T06:54:39Z","description":"room description","name":"room001"},{"created":"2015-08-11T06:54:39Z","description":"room description","name":"room001"}]},"RoomPayload":{"title":"RoomPayload","type":"object","properties":{"created":{"type":"string","description":"Date of creation","example":"1981-01-18T09:52:25Z","format":"date-time"},"description":{"type":"string","description":"description of room","example":"room description","maxLength":400},"name":{"type":"string","description":"Name of room","example":"room001","pattern":"[a-z|0-9]+","minLength":3,"maxLength":20}},"example":{"created":"1981-01-18T09:52:25Z","description":"room description","name":"room001"},"required":["name","description"]},"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}},"responses":{"BadRequest":{"description":"Bad Request"},"NotFound":{"description":"Not Found"},"OK":{"description":"OK","schema":{"$ref":"#/definitions/Account"}},"Unauthorized":{"description":"Unauthorized"}},"securityDefinitions":{"jwt":{"type":"apiKey","description":"\n\n**Security Scopes**:\n  * `api:access`: API access","name":"Authorization","in":"header"}}}